+++ splurge_pub_sub/__init__.py
@@ -54,7 +54,7 @@ from .pubsub import PubSub
 from .types import Callback, MessageData, SubscriberId, Topic
 from .utility import generate_correlation_id, is_valid_correlation_id, validate_correlation_id
 
-__version__ = "2025.2.0"
+__version__ = "2025.3.0"
 __author__ = "Jim Schilling"
 __license__ = "MIT"